Break up the media model
