+++ packages/client-mobile/src/modules/media/model/index.ts
@@ -1,80 +1,9 @@
 import { useCallback, useEffect, useState } from "react";
-import { Platform, PermissionsAndroid } from "react-native";
-import {
-  CameraRoll,
-  PhotoIdentifier,
-} from "@react-native-camera-roll/camera-roll";
+
 import { promptAllowAccess } from "@modules/alert/logic";
 import { usePersistedState } from "@modules/persisted-state/logic";
-
-type MediaItem = PhotoIdentifier["node"];
-
-const platformVersionAndroid = Number(Platform.Version);
-
-const permissions =
-  platformVersionAndroid >= 33
-    ? [
-        PermissionsAndroid.PERMISSIONS.READ_MEDIA_IMAGES,
-        PermissionsAndroid.PERMISSIONS.READ_MEDIA_VIDEO,
-      ]
-    : [PermissionsAndroid.PERMISSIONS.READ_EXTERNAL_STORAGE];
-
-const checkIsPermittedIOS = async () =>
-  !!(await CameraRoll.getPhotos({ first: 1 }));
-
-const checkIsPermittedAndroid = async () => {
-  const promises = permissions.map((permission) =>
-    PermissionsAndroid.check(permission),
-  );
-
-  return Promise.all(promises).then((...results) =>
-    results.every(([result]) => result === true),
-  );
-};
-
-const checkIsPermitted = async () => {
-  try {
-    if (Platform.OS === "ios") {
-      return await checkIsPermittedIOS();
-    } else {
-      return await checkIsPermittedAndroid();
-    }
-  } catch {
-    return false;
-  }
-};
-
-const requestPermissionAndroid = async () => {
-  try {
-    const result = await PermissionsAndroid.requestMultiple(permissions);
-
-    if (platformVersionAndroid >= 33) {
-      const mediaImagesResult = result["android.permission.READ_MEDIA_IMAGES"];
-      const mediaVideoResult = result["android.permission.READ_MEDIA_VIDEO"];
-      return [mediaImagesResult, mediaVideoResult].every(
-        (result) => result === "granted",
-      );
-    } else {
-      return result["android.permission.READ_EXTERNAL_STORAGE"] === "granted";
-    }
-  } catch {
-    return false;
-  }
-};
-
-const requestPermissionIOS = async () => false;
-
-const requestPermission = async () => {
-  try {
-    if (Platform.OS === "ios") {
-      return await requestPermissionIOS();
-    } else {
-      return await requestPermissionAndroid();
-    }
-  } catch {
-    return false;
-  }
-};
+import { MediaItem } from "../types";
+import { checkIsPermitted, getMedia, requestPermission } from "../service";
 
 export const useMediaModel = () => {
   const [isConnected, setIsConnected] = usePersistedState("isConnected", false);
@@ -126,11 +55,7 @@ export const useMediaModel = () => {
   const refresh = useCallback(
     async (params: { createdAfter: Date; createdBefore: Date }) => {
       setIsRefreshing(true);
-      const response = await CameraRoll.getPhotos({
-        first: 100,
-        fromTime: params.createdAfter.getTime(),
-        toTime: params.createdBefore.getTime(),
-      });
+      const response = await getMedia(params);
       const newValue = response.edges.map(
         (photoIdentifier) => photoIdentifier.node,
       );
