+++ packages/client-mobile/src/modules/media/types.ts
@@ -1,0 +1,3 @@
+import { PhotoIdentifier } from "@react-native-camera-roll/camera-roll";
+
+export type MediaItem = PhotoIdentifier["node"];
